Move serialization in their own folder
