+++ packages/rum/src/domain/record/observers/mediaInteractionObserver.ts
@@ -4,7 +4,7 @@ import { NodePrivacyLevel } from '../../../constants'
 import type { MediaInteraction } from '../../../types'
 import { MediaInteractionType } from '../../../types'
 import { getNodePrivacyLevel } from '../privacy'
-import { getSerializedNodeId, hasSerializedNode } from '../serializationUtils'
+import { getSerializedNodeId, hasSerializedNode } from '../serializers'
 import type { ListenerHandler } from './utils'
 import { getEventTarget } from './utils'
