+++ packages/rum/src/domain/record/observers/mouseInteractionObserver.ts
@@ -4,7 +4,7 @@ import { NodePrivacyLevel } from '../../../constants'
 import type { MouseInteraction, MouseInteractionData, BrowserIncrementalSnapshotRecord } from '../../../types'
 import { IncrementalSource, MouseInteractionType } from '../../../types'
 import { getNodePrivacyLevel } from '../privacy'
-import { getSerializedNodeId, hasSerializedNode } from '../serializationUtils'
+import { getSerializedNodeId, hasSerializedNode } from '../serializers'
 import { assembleIncrementalSnapshot } from '../utils'
 import { tryToComputeCoordinates } from './moveObserver'
 import type { ListenerHandler } from './utils'
