+++ packages/rum/src/domain/record/observers/mutationObserver.ts
@@ -15,14 +15,16 @@ import type {
   TextMutation,
 } from '../../../types'
 import { getNodePrivacyLevel, getTextContent } from '../privacy'
-import type { NodeWithSerializedNode } from '../serializationUtils'
+import type { NodeWithSerializedNode } from '../serializers'
 import {
   getElementInputValue,
   getSerializedNodeId,
   hasSerializedNode,
   nodeAndAncestorsHaveSerializedNode,
-} from '../serializationUtils'
-import { serializeNodeWithId, serializeAttribute, SerializationContextStatus } from '../serialize'
+  serializeNodeWithId,
+  serializeAttribute,
+  SerializationContextStatus,
+} from '../serializers'
 import { createMutationBatch } from '../mutationBatch'
 import type { ShadowRootCallBack, ShadowRootsController } from '../shadowRootsController'
