+++ packages/rum/src/domain/record/observers/scrollObserver.ts
@@ -2,7 +2,7 @@ import type { DefaultPrivacyLevel } from '@datadog/browser-core'
 import { DOM_EVENT, throttle, addEventListener } from '@datadog/browser-core'
 import type { ElementsScrollPositions } from '../elementsScrollPositions'
 import { getNodePrivacyLevel } from '../privacy'
-import { getSerializedNodeId, hasSerializedNode } from '../serializationUtils'
+import { getSerializedNodeId, hasSerializedNode } from '../serializers'
 
 import { getScrollX, getScrollY } from '../viewports'
 import type { ScrollPosition } from '../../../types'
