+++ packages/rum/src/domain/record/serializers/index.ts
@@ -1,0 +1,8 @@
+export {
+  getElementInputValue,
+  getSerializedNodeId,
+  hasSerializedNode,
+  nodeAndAncestorsHaveSerializedNode,
+  NodeWithSerializedNode,
+} from './serializationUtils'
+export { serializeDocument, SerializationContextStatus, serializeNodeWithId, serializeAttribute } from './serialize'
